7_Ollama/local-pdf-chatbot/requirements.txt
langchain==0.1.12
langchain-experimental==0.0.54
chromadb==0.4.24
sentence_transformers==2.5.1
pypdf==4.1.0
python-box
streamlit==1.32.2
pyPDF2